+++ CHANGELOG.md
@@ -33,10 +33,13 @@
   `poll-interval` ([#241][241]).
 * battery: module was not thread safe.
 * dwl module reporting only the last part of the title ([#251][251])
+* i3/sway: regression; persistent workspaces shown twice
+  ([#253][253]).
 
 [239]: https://codeberg.org/dnkl/yambar/issues/239
 [241]: https://codeberg.org/dnkl/yambar/issues/241
 [251]: https://codeberg.org/dnkl/yambar/pulls/251
+[253]: https://codeberg.org/dnkl/yambar/issues/253
 
 
 ### Security